Extract table function tests
Move the existing table function test module into its own file.
